REPECTOR :: 폴더구조 변경 - components에 있던 ui를 pages 폴더로 관리하고 Home 컴포넌트를 app.js에서 불러왔습니다
