+++ src/components/Hash/index.js
